requirements.txt
# Core dependencies
pyyaml>=6.0
aiohttp>=3.8.0
boto3>=1.28.0
botocore>=1.31.0
aiosqlite>=0.19.0
jsonschema>=4.17.3
requests>=2.31.0
pydantic>=2.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0 